refactor(stream): reorg types and add docs
Previously, the managers/workers are messed together, which is consuming. In this PR we move them into appropriate files.
