+++ src/stream/src/error.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use risingwave_common::array::ArrayError;
+use risingwave_common::error::tonic::extra::{Score, ScoredError};
 use risingwave_common::secret::SecretError;
 use risingwave_connector::error::ConnectorError;
 use risingwave_expr::ExprError;
@@ -122,3 +123,60 @@ impl From<StreamError> for tonic::Status {
 }
 
 static_assertions::const_assert_eq!(std::mem::size_of::<StreamError>(), 8);
+
+/// A [`StreamError`] with a score, used to find the root cause of actor failures.
+pub type ScoredStreamError = ScoredError<StreamError>;
+
+impl StreamError {
+    /// Score the given error based on hard-coded rules.
+    pub fn with_score(self) -> ScoredStreamError {
+        // Explicitly list all error kinds here to notice developers to update this function when
+        // there are changes in error kinds.
+
+        fn stream_executor_error_score(e: &StreamExecutorError) -> i32 {
+            use crate::executor::error::ErrorKind;
+            match e.inner() {
+                // `ChannelClosed` or `ExchangeChannelClosed` is likely to be caused by actor exit
+                // and not the root cause.
+                ErrorKind::ChannelClosed(_) | ErrorKind::ExchangeChannelClosed(_) => 1,
+
+                // Normal errors.
+                ErrorKind::Uncategorized(_)
+                | ErrorKind::Storage(_)
+                | ErrorKind::ArrayError(_)
+                | ErrorKind::ExprError(_)
+                | ErrorKind::SerdeError(_)
+                | ErrorKind::SinkError(_, _)
+                | ErrorKind::RpcError(_)
+                | ErrorKind::AlignBarrier(_, _)
+                | ErrorKind::ConnectorError(_)
+                | ErrorKind::DmlError(_)
+                | ErrorKind::NotImplemented(_) => 999,
+            }
+        }
+
+        fn stream_error_score(e: &StreamError) -> i32 {
+            use crate::error::ErrorKind;
+            match e.inner() {
+                // `UnexpectedExit` wraps the original error. Score on the inner error.
+                ErrorKind::UnexpectedExit { source, .. } => stream_error_score(source),
+
+                // `BarrierSend` is likely to be caused by actor exit and not the root cause.
+                ErrorKind::BarrierSend { .. } => 1,
+
+                // Executor errors first.
+                ErrorKind::Executor(ee) => 2000 + stream_executor_error_score(ee),
+
+                // Then other errors.
+                ErrorKind::Uncategorized(_)
+                | ErrorKind::Storage(_)
+                | ErrorKind::Expression(_)
+                | ErrorKind::Array(_)
+                | ErrorKind::Secret(_) => 1000,
+            }
+        }
+
+        let score = Score(stream_error_score(&self));
+        ScoredStreamError { error: self, score }
+    }
+}
