+++ src/stream/src/task/barrier_manager/progress.rs
@@ -18,10 +18,9 @@ use std::fmt::{Display, Formatter};
 use risingwave_common::util::epoch::EpochPair;
 use risingwave_pb::stream_service::barrier_complete_response::PbCreateMviewProgress;
 
-use super::LocalBarrierManager;
-use crate::task::ActorId;
 use crate::task::barrier_manager::LocalBarrierEvent::ReportCreateProgress;
-use crate::task::barrier_manager::managed_state::DatabaseManagedBarrierState;
+use crate::task::barrier_worker::managed_state::DatabaseManagedBarrierState;
+use crate::task::{ActorId, LocalBarrierManager};
 
 type ConsumedEpoch = u64;
 type ConsumedRows = u64;
