+++ src/stream/src/task/barrier_worker/tests.rs
